requirements-dev.txt
pytest
pytest-snapshot
testcontainers
testcontainers-neo4j
testcontainers-redis